Port to Python 3
This commit ports the code to run on python 3 by default. Python 2
compatibility is maintained though users will need to run
`/usr/bin/python2 /path/to/mattermost.py` to run it under python 2.

I debated between leaving the default shebang on `/usr/bin/python2`
but python 3 is the future so I personally believe it to be a better
default choice.